Split up VertexTable.ts
