+++ core/frontend/src/render-primitives.ts
@@ -13,9 +13,13 @@ export * from "./render/primitives/mesh/MeshPrimitives";
 
 export * from "./render/primitives/ColorMap";
 export * from "./render/primitives/DisplayParams";
+export * from "./render/primitives/EdgeParams";
 export * from "./render/primitives/PointCloudPrimitive";
+export * from "./render/primitives/PointStringParams";
+export * from "./render/primitives/Polyface";
+export * from "./render/primitives/PolylineParams";
 export * from "./render/primitives/Primitives";
 export * from "./render/primitives/Strokes";
-export * from "./render/primitives/Polyface";
+export * from "./render/primitives/SurfaceParams";
 export * from "./render/primitives/VertexKey";
 export * from "./render/primitives/VertexTable";
