+++ core/frontend/src/render/primitives/VertexTable.ts
@@ -7,14 +7,13 @@
  */
 
 import { assert } from "@itwin/core-bentley";
-import { Point2d, Point3d, Range2d, Vector3d } from "@itwin/core-geometry";
-import {
-  ColorDef, ColorIndex, FeatureIndex, FeatureIndexType, FillFlags, LinePixels, MeshEdge, OctEncodedNormalPair, PolylineData, PolylineTypeFlags,
-  QParams2d, QParams3d, QPoint2d, QPoint3dList, RenderMaterial, RenderTexture,
-} from "@itwin/core-common";
+import { Point2d, Range2d } from "@itwin/core-geometry";
+import { ColorDef, ColorIndex, FeatureIndex, FeatureIndexType, QParams2d, QParams3d, QPoint2d } from "@itwin/core-common";
 import { IModelApp } from "../../IModelApp";
 import { AuxChannelTable } from "./AuxChannelTable";
 import { MeshArgs, PolylineArgs } from "./mesh/MeshPrimitives";
+import { createSurfaceMaterial, SurfaceParams, SurfaceType } from "./SurfaceParams";
+import { EdgeParams } from "./EdgeParams";
 
 /**
  * Holds an array of indices into a VertexTable. Each index is a 24-bit unsigned integer.
@@ -220,480 +219,6 @@ export class VertexTable implements VertexTableProps {
   }
 }
 
-/** Describes point string geometry to be submitted to the rendering system.
- * @internal
- */
-export class PointStringParams {
-  public readonly vertices: VertexTable;
-  public readonly indices: VertexIndices;
-  public readonly weight: number;
-
-  public constructor(vertices: VertexTable, indices: VertexIndices, weight: number) {
-    this.vertices = vertices;
-    this.indices = indices;
-    this.weight = weight;
-  }
-
-  public static create(args: PolylineArgs): PointStringParams | undefined {
-    if (!args.flags.isDisjoint)
-      return undefined;
-
-    const vertices = VertexTable.createForPolylines(args);
-    if (undefined === vertices)
-      return undefined;
-
-    const polylines = args.polylines;
-    let vertIndices = polylines[0].vertIndices;
-    if (1 < polylines.length) {
-      // We used to assert this wouldn't happen - apparently it does...
-      vertIndices = [];
-      for (const polyline of polylines)
-        for (const vertIndex of polyline.vertIndices)
-          vertIndices.push(vertIndex);
-    }
-
-    const vertexIndices = VertexIndices.fromArray(vertIndices);
-    assert(vertexIndices.length === vertIndices.length);
-
-    return new PointStringParams(vertices, vertexIndices, args.width);
-  }
-}
-
-/** Parameter associated with each vertex index of a tesselated polyline. */
-const enum PolylineParam { // eslint-disable-line no-restricted-syntax
-  kNone = 0,
-  kSquare = 1 * 3,
-  kMiter = 2 * 3,
-  kMiterInsideOnly = 3 * 3,
-  kJointBase = 4 * 3,
-  kNegatePerp = 8 * 3,
-  kNegateAlong = 16 * 3,
-  kNoneAdjustWeight = 32 * 3,
-}
-
-/**
- * Represents a tesselated polyline.
- * Given a polyline as a line string, each segment of the line string is triangulated into a quad.
- * Based on the angle between two segments, additional joint triangles may be inserted in between to enable smoothly-rounded corners.
- * @internal
- */
-export interface TesselatedPolyline {
-  /** 24-bit index of each vertex. */
-  readonly indices: VertexIndices;
-  /** 24-bit index of the previous vertex in the polyline. */
-  readonly prevIndices: VertexIndices;
-  /** 24-bit index of the next vertex in the polyline, plus 8-bit parameter describing the semantics of this vertex. */
-  readonly nextIndicesAndParams: Uint8Array;
-}
-
-class PolylineVertex {
-  public isSegmentStart: boolean = false;
-  public isPolylineStartOrEnd: boolean = false;
-  public vertexIndex: number = 0;
-  public prevIndex: number = 0;
-  public nextIndex: number = 0;
-
-  public constructor() { }
-
-  public init(isSegmentStart: boolean, isPolylineStartOrEnd: boolean, vertexIndex: number, prevIndex: number, nextIndex: number) {
-    this.isSegmentStart = isSegmentStart;
-    this.isPolylineStartOrEnd = isPolylineStartOrEnd;
-    this.vertexIndex = vertexIndex;
-    this.prevIndex = prevIndex;
-    this.nextIndex = nextIndex;
-  }
-
-  public computeParam(negatePerp: boolean, adjacentToJoint: boolean = false, joint: boolean = false, noDisplacement: boolean = false): number {
-    if (joint)
-      return PolylineParam.kJointBase;
-
-    let param: PolylineParam;
-    if (noDisplacement)
-      param = PolylineParam.kNoneAdjustWeight; // prevent getting tossed before width adjustment
-    else if (adjacentToJoint)
-      param = PolylineParam.kMiterInsideOnly;
-    else
-      param = this.isPolylineStartOrEnd ? PolylineParam.kSquare : PolylineParam.kMiter;
-
-    let adjust = 0;
-    if (negatePerp)
-      adjust = PolylineParam.kNegatePerp;
-    if (!this.isSegmentStart)
-      adjust += PolylineParam.kNegateAlong;
-
-    return param + adjust;
-  }
-}
-
-class PolylineTesselator {
-  private _polylines: PolylineData[];
-  private _points: QPoint3dList;
-  private _doJoints: boolean;
-  private _numIndices = 0;
-  private _vertIndex: number[] = [];
-  private _prevIndex: number[] = [];
-  private _nextIndex: number[] = [];
-  private _nextParam: number[] = [];
-  private _position: Point3d[] = [];
-
-  public constructor(polylines: PolylineData[], points: QPoint3dList, doJointTriangles: boolean) {
-    this._polylines = polylines;
-    this._points = points;
-    this._doJoints = doJointTriangles;
-  }
-
-  public static fromPolyline(args: PolylineArgs): PolylineTesselator {
-    return new PolylineTesselator(args.polylines, args.points, wantJointTriangles(args.width, args.flags.is2d));
-  }
-
-  public static fromMesh(args: MeshArgs): PolylineTesselator | undefined {
-    if (undefined !== args.edges.polylines.lines && undefined !== args.points)
-      return new PolylineTesselator(args.edges.polylines.lines, args.points, wantJointTriangles(args.edges.width, args.is2d));
-
-    return undefined;
-  }
-
-  public tesselate(): TesselatedPolyline {
-    for (const p of this._points.list)
-      this._position.push(p.unquantize(this._points.params));
-
-    this._tesselate();
-
-    const vertIndex = VertexIndices.fromArray(this._vertIndex);
-    const prevIndex = VertexIndices.fromArray(this._prevIndex);
-
-    const nextIndexAndParam = new Uint8Array(this._numIndices * 4);
-    for (let i = 0; i < this._numIndices; i++) {
-      const index = this._nextIndex[i];
-      const j = i * 4;
-      VertexIndices.encodeIndex(index, nextIndexAndParam, j);
-      nextIndexAndParam[j + 3] = this._nextParam[i] & 0x000000ff;
-    }
-
-    return {
-      indices: vertIndex,
-      prevIndices: prevIndex,
-      nextIndicesAndParams: nextIndexAndParam,
-    };
-  }
-
-  private _tesselate() {
-    const v0 = new PolylineVertex(), v1 = new PolylineVertex();
-    const maxJointDot = -0.7;
-
-    for (const line of this._polylines) {
-      if (line.numIndices < 2)
-        continue;
-
-      const last = line.numIndices - 1;
-      const isClosed: boolean = line.vertIndices[0] === line.vertIndices[last];
-
-      for (let i = 0; i < last; ++i) {
-        const idx0 = line.vertIndices[i];
-        const idx1 = line.vertIndices[i + 1];
-        const isStart: boolean = (0 === i);
-        const isEnd: boolean = (last - 1 === i);
-        const prevIdx0 = isStart ? (isClosed ? line.vertIndices[last - 1] : idx0) : line.vertIndices[i - 1];
-        const nextIdx1 = isEnd ? (isClosed ? line.vertIndices[1] : idx1) : line.vertIndices[i + 2];
-
-        v0.init(true, isStart && !isClosed, idx0, prevIdx0, idx1);
-        v1.init(false, isEnd && !isClosed, idx1, nextIdx1, idx0);
-
-        const jointAt0: boolean = this._doJoints && (isClosed || !isStart) && this._dotProduct(v0) > maxJointDot;
-        const jointAt1: boolean = this._doJoints && (isClosed || !isEnd) && this._dotProduct(v1) > maxJointDot;
-
-        if (jointAt0 || jointAt1) {
-          this._addVertex(v0, v0.computeParam(true, jointAt0, false, false));
-          this._addVertex(v1, v1.computeParam(false, jointAt1, false, false));
-          this._addVertex(v0, v0.computeParam(false, jointAt0, false, true));
-          this._addVertex(v0, v0.computeParam(false, jointAt0, false, true));
-          this._addVertex(v1, v1.computeParam(false, jointAt1, false, false));
-          this._addVertex(v1, v1.computeParam(false, jointAt1, false, true));
-          this._addVertex(v0, v0.computeParam(false, jointAt0, false, true));
-          this._addVertex(v1, v1.computeParam(false, jointAt1, false, true));
-          this._addVertex(v0, v0.computeParam(false, jointAt0, false, false));
-          this._addVertex(v0, v0.computeParam(false, jointAt0, false, false));
-          this._addVertex(v1, v1.computeParam(false, jointAt1, false, true));
-          this._addVertex(v1, v1.computeParam(true, jointAt1, false, false));
-
-          if (jointAt0)
-            this.addJointTriangles(v0, v0.computeParam(false, true, false, true), v0);
-
-          if (jointAt1)
-            this.addJointTriangles(v1, v1.computeParam(false, true, false, true), v1);
-        } else {
-          this._addVertex(v0, v0.computeParam(true));
-          this._addVertex(v1, v1.computeParam(false));
-          this._addVertex(v0, v0.computeParam(false));
-          this._addVertex(v0, v0.computeParam(false));
-          this._addVertex(v1, v1.computeParam(false));
-          this._addVertex(v1, v1.computeParam(true));
-        }
-      }
-    }
-  }
-
-  private addJointTriangles(v0: PolylineVertex, p0: number, v1: PolylineVertex): void {
-    const param = v1.computeParam(false, false, true);
-    for (let i = 0; i < 3; i++) {
-      this._addVertex(v0, p0);
-      this._addVertex(v1, param + i + 1);
-      this._addVertex(v1, param + i);
-    }
-  }
-
-  private _dotProduct(v: PolylineVertex): number {
-    const pos: Point3d = this._position[v.vertexIndex];
-    const prevDir: Vector3d = Vector3d.createStartEnd(this._position[v.prevIndex], pos);
-    const nextDir: Vector3d = Vector3d.createStartEnd(this._position[v.nextIndex], pos);
-    return prevDir.dotProduct(nextDir);
-  }
-
-  private _addVertex(vertex: PolylineVertex, param: number): void {
-    this._vertIndex[this._numIndices] = vertex.vertexIndex;
-    this._prevIndex[this._numIndices] = vertex.prevIndex;
-    this._nextIndex[this._numIndices] = vertex.nextIndex;
-    this._nextParam[this._numIndices] = param;
-    this._numIndices++;
-  }
-}
-
-/** Strictly for tests. @internal */
-export function tesselatePolyline(polylines: PolylineData[], points: QPoint3dList, doJointTriangles: boolean): TesselatedPolyline {
-  const tesselator = new PolylineTesselator(polylines, points, doJointTriangles);
-  return tesselator.tesselate();
-}
-
-/** @internal */
-export enum SurfaceType {
-  Unlit,
-  Lit,
-  Textured,
-  TexturedLit,
-  VolumeClassifier,
-}
-
-/** @internal */
-export function isValidSurfaceType(value: number): boolean {
-  switch (value) {
-    case SurfaceType.Unlit:
-    case SurfaceType.Lit:
-    case SurfaceType.Textured:
-    case SurfaceType.TexturedLit:
-    case SurfaceType.VolumeClassifier:
-      return true;
-    default:
-      return false;
-  }
-}
-
-/** @internal */
-export interface SurfaceRenderMaterial {
-  readonly isAtlas: false;
-  readonly material: RenderMaterial;
-}
-
-/** @internal */
-export interface SurfaceMaterialAtlas {
-  readonly isAtlas: true;
-  // Overrides surface alpha to be translucent. Implies `overridesAlpha`.
-  readonly hasTranslucency: boolean;
-  // Overrides surface alpha to be opaque or translucent.
-  readonly overridesAlpha: boolean;
-  // offset past the END of the vertex data; equivalently, number of 32-bit colors in color table preceding material atlas.
-  readonly vertexTableOffset: number;
-  readonly numMaterials: number;
-}
-
-/** @internal */
-export type SurfaceMaterial = SurfaceRenderMaterial | SurfaceMaterialAtlas;
-
-/** @internal */
-export function createSurfaceMaterial(source: RenderMaterial | undefined): SurfaceMaterial | undefined {
-  if (undefined === source)
-    return undefined;
-  else
-    return { isAtlas: false, material: source };
-}
-
-/** @internal */
-export interface SurfaceParams {
-  readonly type: SurfaceType;
-  readonly indices: VertexIndices;
-  readonly fillFlags: FillFlags;
-  readonly hasBakedLighting: boolean;
-  readonly hasFixedNormals: boolean;
-  readonly textureMapping?: {
-    texture: RenderTexture;
-    alwaysDisplayed: boolean;
-  };
-  readonly material?: SurfaceMaterial;
-}
-
-/**
- * Describes a set of line segments representing edges of a mesh.
- * Each segment is expanded into a quad defined by two triangles.
- * The positions are adjusted in the shader to account for the edge width.
- * @internal
- */
-export interface SegmentEdgeParams {
-  /** The 24-bit indices of the tesselated line segment */
-  readonly indices: VertexIndices;
-  /**
-   * For each 24-bit index, 4 bytes:
-   * the 24-bit index of the vertex at the other end of the segment, followed by
-   * an 8-bit 'quad index' in [0..3] indicating which point in the expanded quad the vertex represents.
-   */
-  readonly endPointAndQuadIndices: Uint8Array;
-}
-
-function convertPolylinesAndEdges(polylines?: PolylineData[], edges?: MeshEdge[]): SegmentEdgeParams | undefined {
-  let numIndices = undefined !== edges ? edges.length : 0;
-  if (undefined !== polylines)
-    for (const pd of polylines)
-      numIndices += (pd.vertIndices.length - 1);
-
-  if (0 === numIndices)
-    return undefined;
-
-  numIndices *= 6;
-  const indexBytes = new Uint8Array(numIndices * 3);
-  const endPointAndQuadIndexBytes = new Uint8Array(numIndices * 4);
-
-  let ndx: number = 0;
-  let ndx2: number = 0;
-
-  const addPoint = (p0: number, p1: number, quadIndex: number) => {
-    VertexIndices.encodeIndex(p0, indexBytes, ndx);
-    ndx += 3;
-    VertexIndices.encodeIndex(p1, endPointAndQuadIndexBytes, ndx2);
-    endPointAndQuadIndexBytes[ndx2 + 3] = quadIndex;
-    ndx2 += 4;
-  };
-
-  if (undefined !== polylines) {
-    for (const pd of polylines) {
-      const num = pd.vertIndices.length - 1;
-      for (let i = 0; i < num; ++i) {
-        let p0 = pd.vertIndices[i];
-        let p1 = pd.vertIndices[i + 1];
-        if (p1 < p0) { // swap so that lower index is first.
-          p0 = p1;
-          p1 = pd.vertIndices[i];
-        }
-        addPoint(p0, p1, 0);
-        addPoint(p1, p0, 2);
-        addPoint(p0, p1, 1);
-        addPoint(p0, p1, 1);
-        addPoint(p1, p0, 2);
-        addPoint(p1, p0, 3);
-      }
-    }
-  }
-
-  if (undefined !== edges) {
-    for (const meshEdge of edges) {
-      const p0 = meshEdge.indices[0];
-      const p1 = meshEdge.indices[1];
-      addPoint(p0, p1, 0);
-      addPoint(p1, p0, 2);
-      addPoint(p0, p1, 1);
-      addPoint(p0, p1, 1);
-      addPoint(p1, p0, 2);
-      addPoint(p1, p0, 3);
-    }
-  }
-
-  return {
-    indices: new VertexIndices(indexBytes),
-    endPointAndQuadIndices: endPointAndQuadIndexBytes,
-  };
-}
-
-/**
- * A set of line segments representing edges of curved portions of a mesh.
- * Each vertex is augmented with a pair of oct-encoded normals used in the shader
- * to determine whether or not the edge should be displayed.
- * @internal
- */
-export interface SilhouetteParams extends SegmentEdgeParams {
-  /** Per index, 2 16-bit oct-encoded normals */
-  readonly normalPairs: Uint8Array;
-}
-
-function convertSilhouettes(edges: MeshEdge[], normalPairs: OctEncodedNormalPair[]): SilhouetteParams | undefined {
-  const base = convertPolylinesAndEdges(undefined, edges);
-  if (undefined === base)
-    return undefined;
-
-  const normalPairBytes = new Uint8Array(normalPairs.length * 6 * 4);
-  const normalPair16 = new Uint16Array(normalPairBytes.buffer);
-
-  let ndx = 0;
-  for (const pair of normalPairs) {
-    for (let i = 0; i < 6; i++) {
-      normalPair16[ndx++] = pair.first.value;
-      normalPair16[ndx++] = pair.second.value;
-    }
-  }
-
-  return {
-    indices: base.indices,
-    endPointAndQuadIndices: base.endPointAndQuadIndices,
-    normalPairs: normalPairBytes,
-  };
-}
-
-/** Describes the edges of a mesh. */
-export interface EdgeParams {
-  /** The edge width in pixels. */
-  readonly weight: number;
-  /** The line pattern in which edges are drawn. */
-  readonly linePixels: LinePixels;
-  /** Simple single-segment edges, always displayed when edge display is enabled. */
-  readonly segments?: SegmentEdgeParams;
-  /** Single-segment edges of curved surfaces, displayed based on edge normal relative to eye. */
-  readonly silhouettes?: SilhouetteParams;
-  /** Polyline edges, always displayed when edge display is enabled. */
-  readonly polylines?: TesselatedPolyline;
-}
-
-function wantJointTriangles(weight: number, is2d: boolean): boolean {
-  // Joints are incredibly expensive. In 3d, only generate them if the line is sufficiently wide for them to be noticeable.
-  const jointWidthThreshold = 3;
-  return is2d || weight >= jointWidthThreshold;
-}
-
-function convertEdges(meshArgs: MeshArgs): EdgeParams | undefined {
-  const args = meshArgs.edges;
-  if (undefined === args)
-    return undefined;
-
-  let polylines: TesselatedPolyline | undefined;
-  let segments: SegmentEdgeParams | undefined;
-  if (wantJointTriangles(args.width, meshArgs.is2d)) {
-    segments = convertPolylinesAndEdges(args.polylines.lines, args.edges.edges);
-  } else {
-    segments = convertPolylinesAndEdges(undefined, args.edges.edges);
-    const tesselator = PolylineTesselator.fromMesh(meshArgs);
-    if (undefined !== tesselator)
-      polylines = tesselator.tesselate();
-  }
-
-  // ###TODO: why the heck are the edges and normals of SilhouetteEdgeArgs potentially undefined???
-  const silhouettes = undefined !== args.silhouettes.edges && undefined !== args.silhouettes.normals ? convertSilhouettes(args.silhouettes.edges, args.silhouettes.normals) : undefined;
-  if (undefined === segments && undefined === silhouettes && undefined === polylines)
-    return undefined;
-
-  return {
-    weight: args.width,
-    linePixels: args.linePixels,
-    segments,
-    silhouettes,
-    polylines,
-  };
-}
-
 /**
  * Describes mesh geometry to be submitted to the rendering system.
  * A mesh consists of a surface and its edges, which may include any combination of silhouettes, polylines, and single segments.
@@ -733,46 +258,8 @@ export class MeshParams {
     };
 
     const channels = undefined !== args.auxChannels ? AuxChannelTable.fromChannels(args.auxChannels, vertices.numVertices) : undefined;
-    const edges = convertEdges(args);
+    const edges = EdgeParams.fromMeshArgs(args);
     return new MeshParams(vertices, surface, edges, args.isPlanar, channels);
-  }
-}
-
-/**
- * Describes a set of tesselated polylines.
- * Each segment of each polyline is triangulated into a quad. Additional triangles may be inserted
- * between segments to enable rounded corners.
- */
-export class PolylineParams {
-  public readonly vertices: VertexTable;
-  public readonly polyline: TesselatedPolyline;
-  public readonly isPlanar: boolean;
-  public readonly type: PolylineTypeFlags;
-  public readonly weight: number;
-  public readonly linePixels: LinePixels;
-
-  /** Directly construct a PolylineParams. The PolylineParams takes ownership of all input data. */
-  public constructor(vertices: VertexTable, polyline: TesselatedPolyline, weight: number, linePixels: LinePixels, isPlanar: boolean, type: PolylineTypeFlags = PolylineTypeFlags.Normal) {
-    this.vertices = vertices;
-    this.polyline = polyline;
-    this.isPlanar = isPlanar;
-    this.weight = weight;
-    this.linePixels = linePixels;
-    this.type = type;
-  }
-
-  /** Construct from an PolylineArgs. */
-  public static create(args: PolylineArgs): PolylineParams | undefined {
-    assert(!args.flags.isDisjoint);
-    const vertices = VertexTable.createForPolylines(args);
-    if (undefined === vertices)
-      return undefined;
-
-    const tesselator = PolylineTesselator.fromPolyline(args);
-    if (undefined === tesselator)
-      return undefined;
-
-    return new PolylineParams(vertices, tesselator.tesselate(), args.width, args.linePixels, args.flags.isPlanar, args.flags.type);
   }
 }
