+++ core/frontend/src/render/webgl/CachedGeometry.ts
@@ -11,7 +11,7 @@ import { Angle, Point2d, Point3d, Range3d, Vector2d, Vector3d } from "@itwin/cor
 import { Npc, QParams2d, QParams3d, QPoint2dList, QPoint3dList, RenderMode, RenderTexture } from "@itwin/core-common";
 import { RenderSkyGradientParams, RenderSkySphereParams } from "../RenderSystem";
 import { FlashMode } from "../../FlashSettings";
-import { TesselatedPolyline } from "../primitives/VertexTable";
+import { TesselatedPolyline } from "../primitives/PolylineParams";
 import { RenderMemory } from "../RenderMemory";
 import { AttributeMap } from "./AttributeMap";
 import { ColorInfo } from "./ColorInfo";
