+++ core/frontend/src/render/webgl/EdgeGeometry.ts
@@ -9,7 +9,8 @@
 import { MeshData, MeshGeometry } from "./Mesh";
 import { assert, dispose } from "@itwin/core-bentley";
 import { RenderMode } from "@itwin/core-common";
-import { SegmentEdgeParams, SilhouetteParams, TesselatedPolyline } from "../primitives/VertexTable";
+import { TesselatedPolyline } from "../primitives/PolylineParams";
+import { SegmentEdgeParams, SilhouetteParams } from "../primitives/EdgeParams";
 import { RenderMemory } from "../RenderMemory";
 import { AttributeMap } from "./AttributeMap";
 import { PolylineBuffers } from "./CachedGeometry";
