+++ core/frontend/src/render/webgl/Mesh.ts
@@ -10,7 +10,8 @@ import { assert, dispose } from "@itwin/core-bentley";
 import { Point3d, Range3d } from "@itwin/core-geometry";
 import { FeatureIndexType, FillFlags, LinePixels, RenderMode } from "@itwin/core-common";
 import { InstancedGraphicParams } from "../InstancedGraphicParams";
-import { MeshParams, SurfaceType } from "../primitives/VertexTable";
+import { MeshParams } from "../primitives/VertexTable";
+import { SurfaceType } from "../primitives/SurfaceParams";
 import { RenderMemory } from "../RenderMemory";
 import { RenderGeometry } from "../RenderSystem";
 import { CachedGeometry, LUTGeometry } from "./CachedGeometry";
