+++ core/frontend/src/test/render/GraphicBuilder.test.ts
@@ -13,7 +13,8 @@ import { IModelConnection } from "../../IModelConnection";
 import { createBlankConnection } from "../createBlankConnection";
 import { RenderSystem } from "../../render/RenderSystem";
 import { ScreenViewport } from "../../Viewport";
-import { MeshParams, SurfaceType } from "../../render/primitives/VertexTable";
+import { MeshParams } from "../../render/primitives/VertexTable";
+import { SurfaceType } from "../../render/primitives/SurfaceParams";
 import { MeshArgs } from "../../render/primitives/mesh/MeshPrimitives";
 import { MeshGraphic } from "../../render/webgl/Mesh";
 import { InstancedGraphicParams } from "../../render/InstancedGraphicParams";
